toReadableTime in external file and bumped package
