+++ index.js
@@ -4,6 +4,7 @@ var log
 
 var icons = require('./assets/icons')
 var io = require('socket.io-client')
+var utilities = require('./utilities')
 var socket = null
 
 let timer = {}
@@ -63,10 +64,10 @@ instance.prototype.initModule = function () {
     })
 
     socket.on('timer', (data) => {
-      //self.log('info', JSON.stringify(data))
       timer = data
 
-      readable = self.toReadableTime(timer.running)
+      let readable = utilities.toReadableTime(timer.running, timer.isNegative)
+
       self.setVariables({
         time: readable.hours + ':' + readable.minutes + ':' + readable.seconds,
         time_hm: readable.hours + ':' + readable.minutes,
@@ -124,44 +125,6 @@ instance.prototype.config_fields = function () {
       required: true,
     },
   ]
-}
-
-instance.prototype.toReadableTime = function (secondsfrommidnight) {
-  const SECONDS_IN_DAY = 86400
-
-  secondsfrommidnight = Number(secondsfrommidnight) % SECONDS_IN_DAY
-  if (secondsfrommidnight < 0) {
-    secondsfrommidnight = secondsfrommidnight * -1
-  }
-
-  let hours = Math.floor(secondsfrommidnight / 60 / 60)
-  let minutes = Math.floor((secondsfrommidnight - hours * 60 * 60) / 60)
-  let seconds = Math.floor(secondsfrommidnight - hours * 60 * 60 - minutes * 60)
-
-  // deal with hours
-  if (hours < 10) {
-    hours = '0' + hours
-  }
-
-  if (timer.isNegative) {
-    hours = '-' + hours
-  }
-
-  if (minutes < 10) {
-    // deal with minutes
-    minutes = '0' + minutes
-  }
-
-  // deal with seconds
-  if (seconds < 10) {
-    seconds = '0' + seconds
-  }
-
-  return {
-    hours: String(hours),
-    minutes: String(minutes),
-    seconds: String(seconds),
-  }
 }
 
 instance.prototype.actions = function (system) {
